refactor: move banner scripts into banner context
